fallback when creating symlinks fails
added a fallback when creating symbolic links to images while creating bundler output. If symbolic link fails, images are copied
